Removed output directories in case that messes with the testing.
I'd prefer the output files go into somewhere other than the same
directory as the main script. The instructions didn't specify it had to
be in the same dir so I initially wrote them elsewhere. Figuring most
people didn't do this and that a script might be used for testing, I
dropped that functionality and am just writing them next to the script.